z_iptables: Set DNS also during 'blocking run', if defined
Normally, blocking and setting a custom DNS are own services
independent from each other. However - if both properties are
set and the services are triggered at the same time, e.g. during
boot, the running instances of iptables/ip6tables to set the
blocking rules may 'block' the setting of the DNAT rules.

In order to work-around this situation, the DNS will now also
be set according to the existing DNS property value, when the
blocking takes place.

Change-Id: I90222170a1e2a8e52aaaf893ffe400c13e77785c